Refactor reusable proxy into trait.
